Refactor DocumentProcessorTest
I've externalized the SimpleEnglish results to a properties file,
and added a method to read it into a collection.

Also removed the sorting stuff, because it was dependent on the
WordOccurence compareTo method. Now it works on sets, and they're
compared by a Hamcrest expression.

Since Hamcrest produces useful error messages, manual listing via
System.out turned out to be unnecessary as well.