Alter CsvExportWords, create test class
CsvExportWords and the corresponding interface have been modified
a bit so that they can be tested properly. Now the class relies on
constructor injections and needs to be closed manually after use.

Also created a test for CsvExportWords, which checks the output
and file closing. Since it uses Mockito for mocking purposes,
it has been added as a dependency.